Numerous improvements
 * Data and Mesh option widgets now sync with loads/runs
 * Options are unavailable when the current data is insufficient
 * Adaptive vs. Constant is working
 * Face / Cell locks work independently
 * Toggle on/off data widget view
 * Removed useless camera options
 * Only 1 sizing field, 1 volume, 1 mesh, and > 1 indicators allowed

+++ src/lib/cleaver/Volume.cpp
@@ -165,7 +165,7 @@ double Volume::valueAt(double x, double y, double z, int material) const
 
 int Volume::numberOfMaterials() const
 {
-    return m_valueFields.size();
+    return static_cast<int>(m_valueFields.size());
 }
 
 const BoundingBox& Volume::bounds() const
